End Day

+++ Assets/Scripts/Model/Component/BuildingExtractionOnCellC.cs
@@ -2,8 +2,8 @@
 {
     public sealed class BuildingExtractionOnCellC
     {
-        public float HowManyWoodcutterCanExtractWood { get; internal set; }
-        public float HowManyFarmCanExtractFood { get; internal set; }
+        public double HowManyWoodcutterCanExtractWood { get; internal set; }
+        public double HowManyFarmCanExtractFood { get; internal set; }
 
         public bool CanWoodcutterExtact => HowManyWoodcutterCanExtractWood > 0;
         public bool CanFarmExtact => HowManyFarmCanExtractFood > 0;
