+++ Assets/Scripts/Model/Entity/EnvironmentE.cs
@@ -4,7 +4,7 @@ namespace Chessy.Model
 {
     public sealed class EnvironmentE
     {
-        public readonly EnvironmentC EnvironmentC = new(new float[(byte)EnvironmentTypes.End]);
+        public readonly EnvironmentC EnvironmentC = new(new double[(byte)EnvironmentTypes.End]);
 
         internal void Dispose()
         {
